LayoutTests/platform/ios-sim-deprecated/http/tests/security/contentTypeOptions/nosniff-script-blocked-expected.txt

CONSOLE MESSAGE: line 1: Executed script with MIME type: 'application/json'.
CONSOLE MESSAGE: line 1: Executed script with MIME type: 'image/png'.
CONSOLE MESSAGE: line 1: Executed script with MIME type: 'text/html'.
CONSOLE MESSAGE: line 1: Executed script with MIME type: 'text/vbs'.
CONSOLE MESSAGE: line 1: Executed script with MIME type: 'text/vbscript'.
CONSOLE MESSAGE: line 1: Executed script with MIME type: 'text/x-javascript'.
Check that script sent with an 'X-Content-Type-Options: nosniff' header is correctly blocked if the MIME type isn't scripty.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL window.scriptsSuccessfullyLoaded should be 0. Was 6.
PASS successfullyParsed is true

TEST COMPLETE
